requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
typing-extensions==4.12.2
email-validator==2.1.0.post1
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.24.1
pandas==2.2.3
numpy>=2.1.0
openpyxl==3.1.2
reportlab==4.0.7
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
